front: controllers: Optimize getGeometry function using ST_Simplify in PostGIS query.
This commit enhances the performance of the getGeometry function by
incorporating the ST_Simplify function into the SQL query. The geometry
simplification is done with a tolerance level of 0.01, which should be adjusted
based on the specific needs of the application.

Note: The tolerance level for ST_Simplify may need further tuning.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>