api: Add optional 'resolveEmpty' query parameter for debug and optimization purposes.
Added a new optional query parameter 'resolveEmpty' to the API's geometry
endpoint. This parameter allows the caller to specify whether to resolve empty
geometries by aggregating the geometries of the subregions. The parameter is of
boolean type and is not required. This addition is particularly useful for
debugging and optimization, as it allows avoiding the retrieval of geometry for
large regions when not necessary.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>