deploy: prioritize country levels and skip non-essential ones in init-regions-table.py.
Prioritize country levels by adhering to an order of 'COUNTRY', 'GOVERNEDBY',
and 'SOVEREIGN'. This ensures that only the highest-priority country level is
used for each feature. This change helps in reducing redundancy and focuses on
the most relevant geographical levels.

Example:
Before: Both 'France' and 'Saint-Barthélemy' would be treated as separate
country levels for the Saint-Barthélemy feature.
After: Only 'Saint-Barthélemy' is used, as it is the highest-priority country
level for this feature.

Added a new dictionary 'identified_country_levels' to keep track of the chosen
country level for each feature.

The change improves the accuracy and relevancy of the regions table.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>