deploy: Fix gadm_uid population in regions table.
This commit addresses a bug where the gadm_uid field in the regions table was
not being populated correctly, and was almost always NULL. The issue was due to
the uid variable being set to None whenever a region with subregions was
encountered, and not being updated again before reaching a region without
subregions. By moving the assignment of uid to inside the else block, we ensure
that uid is updated with the correct value from the current feature whenever a
region without subregions is encountered. This change resolves the issue and
ensures correct population of the gadm_uid field in the regions table.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>