deploy: Refactor region hierarchy and redefine level priorities.
1. Reordered predefined geographical levels to prioritize 'SOVEREIGN' entities
over 'COUNTRY'. This change affects the organization and display of regions,
making it more logical for cases like UK territories.

2. Fixed issues with duplicate or unnecessary levels in the region hierarchy.
Added logic to skip non-prioritized subcountry levels and prevent duplication
of country or region names.

These changes aim to improve the accuracy and readability of the generated
region hierarchies.

Issue: #42, #29

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>