deploy: Update README to reflect improved DB initialization process and .env handling.
Enhanced the database initialization documentation in the README to include:

- A more detailed explanation of the automated DB initialization process using
  the GADM file.
- The addition of hierarchical data processing and indexing for better
  performance.
- New options to toggle the inclusion of geometric data to accelerate the
  initial setup.

Expanded the environment configuration files section to detail the precedence
and usage of various `.env*` files, emphasizing the importance of `.env.local`
for sensitive overrides.

Added a comprehensive section for `init-regions-table` script parameters,
explaining the `--fast` and `--geometry` options and their impact on the
database initialization process within Docker. Also provided instructions on
setting these parameters in the `init-db.sh` script, ensuring users understand
how to customize their initialization process.

Included a note on the importance of rebuilding the Docker image after changes
to the `init-db.sh` script for the updated initialization parameters to take
effect.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>