deploy: Add validation for GADM file argument in init-regions script.
This commit introduces an additional validation check in the
init-regions-table.py script. It ensures that if the 'alt_only' flag is not
set, the path to the GADM GeoPackage file must be provided using the -s option.
If this argument is missing, the script will now display an error message and
terminate.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>