back: Add @turf/turf Dependency.
This commit introduces the addition of the `@turf/turf` library to the
project's backend dependencies. The specific version added is `6.5.0`. This
library is expected to enhance the backend's geographical data processing
capabilities, providing advanced geospatial analysis functionalities that are
critical for the project's requirements related to region handling and geometry
calculations.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>