front: Add ESLint configuration and lint script to package.json.
This commit introduces ESLint as a development dependency in the project, with
the specific version set to ^8.54.0. Additionally, a new script 'lint' has been
added to the scripts section. This script enables code linting for .js and .jsx
files within the 'src' directory, ensuring adherence to coding standards and
improving code quality.

Issue: #98

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>