back: Replace ++ with += 1 in regionController for loop.
Modified the for loop in the 'getGeometry' function of regionController.js to
use 'i += 1' instead of 'i++' for incrementation. This change aligns with the
ESLint 'no-plusplus' rule, enhancing code readability and maintaining
consistency with our codebase's style guidelines. The functionality of the loop
remains unchanged while adhering to the preferred coding standards.

Issue: #98

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>