Add GitHub Action for ESLint checks on backend code.
This commit adds a new GitHub Actions workflow, defined in 'eslint-back.yml',
to perform ESLint checks on the backend code. This workflow activates on pull
requests targeting the main branch, specifically when there are changes in
JavaScript files under the 'backend' directory. It runs on the latest Ubuntu
environment and performs several steps: checking out the code, setting up
Node.js (version 20), installing required dependencies in the 'backend'
directory, and finally executing the ESLint check using the 'npm run lint'
command. This enhancement ensures adherence to coding standards and maintains
code quality in the backend part of the project.

Issue: #98

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>