front: Enhance Map Initialization Safety in RegionMap Component.
Implemented an additional safeguard in the RegionMap component to verify the
presence of mapContainer.current before initiating the map setup. This
preventive measure is aimed at addressing potential issues where the map
initialization process might encounter a null or undefined mapContainer, such
as in scenarios of component unmounting or incomplete DOM element references.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>