back: Eliminate vulnerable Inflight dependencies by overriding glob version.
Addressed a security vulnerability linked to older versions of the `glob`
package, as discussed in https://github.com/isaacs/inflight/issues/18.
By setting the `glob` version to `10.3.10` in the overrides section of backend
`package.json` file, this commit resolves the issue by ensuring a safer
updated version of `glob` is used throughout the project.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>