Rename GitHub Actions jobs for clear distinction in branch protection.
To resolve the ambiguity in branch protection settings due to having identical
job names in both the backend and frontend ESLint workflows, this commit
renames the jobs. The job in the eslint-back.yml workflow is renamed from
'lint' to 'lint-backend', and the job in the eslint-front.yml workflow is
renamed from 'lint' to 'lint-frontend'. This change ensures that each
workflow's checks are distinctly recognized and can be independently specified
in the branch protection rules.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>