Implement distinct caching for node modules in front and back workflows.
This commit introduces separate caching mechanisms for the node modules in the
frontend and backend workflows of our GitHub Actions. By caching the
node_modules directory in both workflows, we aim to improve the efficiency and
speed of our CI/CD process. The key for the cache in the backend workflow is
based on the OS of the runner and the hash of 'backend/package-lock.json',
while the frontend workflow uses the hash of 'frontend/package-lock.json'. This
distinction ensures that each workflow utilizes its specific set of
dependencies, avoiding potential conflicts or incorrect cache usage. The use of
'actions/cache@v3' facilitates this process, allowing for faster retrieval and
installation of dependencies in subsequent workflow runs.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>