front: Prevent redundant region selection.
This commit introduces a check in the `handleItemClick` function within the
`ListOfRegions` component. The added condition prevents the reselection of an
already selected region. This change enhances the user experience by
eliminating unnecessary re-rendering and potential network requests when a user
clicks on the region that is currently selected.

Issue: #103

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>