front: Optimize map initialization for existing instances.
This commit introduces an optimization in the `MapComponent` within the
`RegionMap` component. The `initializeMap` function now checks if a map
instance already exists. If so, it updates the polygon data source and fits the
map bounds to the new data, instead of creating a new map instance. This
enhancement reduces unnecessary re-initializations of the map, improving
performance and user experience, particularly when updating the displayed
region.

Issue: #113

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>