deploy: Update README for Prepopulated DB Container.
This commit updates the README to reflect recent changes in the project
configuration and processes. Key changes include:

The modification in environment variable setup: The instructions now specify
copying `.env.development.example` to `.env.development`, including the setup
for GitHub Container Registry (GHCR) credentials and PostGIS version.

Introduction of a new section on Prepopulated Database Container: This addition
outlines the process for creating and using a prepopulated database container,
providing guidance on how to build, push, and start the container using the
Makefile. It also includes details on specific Makefile flags and additional
targets related to the prepopulated database container.

Issue: #101

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>