deploy: Add new indices and pg_trgm extension to database initialization script.
Enhanced the database initialization script for the regions table by adding
several key indices and the pg_trgm extension. The script now includes the
creation of indices for `hierarchy_id` and `parent_id` in the hierarchy table,
and a GIN index on `region_name` using `pg_trgm`. These additions aim to
optimize query performance, particularly for text searches and hierarchical
data processing. The inclusion of the pg_trgm extension is essential for
efficient similarity searches in the database. This update will significantly
improve the performance and scalability of database operations related to the
regions table.

Issue: #150

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>