front: Allow force search on Enter Key press.
Modified the Search component to allow force searching by pressing the Enter
key. This change was implemented by creating a new state variable 'searchTerm'
as an object with 'name' and 'force' properties. The 'force' property is set to
true when the Enter key is pressed, triggering the fetchResults function to
execute immediately rather than waiting for the user to stop typing.

Issue: #151

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>