front: Refine Search Logic and Improve Option String Handling.
The Search component has been updated to check if the selected region is
the same as the search term, and if so, the function will return early.
This change prevents unnecessary searches when the selected region and
search term are identical. Additionally, the commit improves the
function that transforms an option into a string representation, making
it more robust. If the option is a string, it is returned directly. If
it is not an object or does not have a name property, an empty string is
returned. Lastly, the unnecessary console log in the onInputChange event
handler has been removed.

Issue: #151

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>