back: Refactor region search query to prevent false positives in security analysis.
Refactored the region search functionality in the backend by changing the way
CASE statements are generated in the SQL query. Previously, the CASE statements
for regex and substring matches were created using template literals with map.
This approach was flagged by semgrep as a potential Sequelize injection point.
The refactor introduces two separate arrays (regexMatchCaseStatements and
substringMatchCaseStatements), which are populated using a for loop and then
joined with SQL's '+' operator. This change simplifies the query construction
and addresses the false positive reported by semgrep, improving the security
analysis of the code.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>