front: Fix Dynamic Update of Autocomplete Dropdown in Search Component.
Fixed an issue where the autocomplete dropdown in the Search component was not
updating correctly with user input. The dropdown now dynamically updates with
search results as the user types, reflecting successful backend API responses.
This fix involved disabling the built-in filtering of the Autocomplete
component, as per MUI documentation guidelines, to allow for better handling of
dynamic search results.

Issue: #171

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>