back: Enhance search query handling with special characters.
Added lodash to the project dependencies and utilized its escapeRegExp function
to escape special characters in the search query. This update is to ensure that
when special characters are included in search terms, they are treated as
literal characters in the regex used in the searchRegions function. These
changes aim to prevent errors and crashes in the backend when users input
special characters like '?' in the search field.

Issue: #160

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>