front: Prevent re-rendering of rendered regions in RegionMap.
Enhanced the RegionMap component by implementing a mechanism to avoid
re-rendering regions that are already rendered. This update introduces state
management for tracking rendered features, coupled with a function
`updateSelectedRegionStyle` that selectively updates the style of a newly
selected region if it's already present in the rendered features array. This
change significantly reduces the computational overhead by preventing redundant
re-rendering of the map, thereby optimizing the performance and responsiveness
of the map interface.

Issue: #185

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>